// modules/apps/headless/headless-admin-site/headless-admin-site-client-js/src/models/IconImageURL.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */


/**
 * @author Rubén Pulido
 * @generated
 */

	/**
	* An Icon Image URL.
	*/
	export class IconImageURL {
			"url"?: string;

		static "discriminator": string | undefined = undefined;

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "url",
			name: "url",
			type: "string",
		},
		];

		static getAttributeTypeMap() {
				return IconImageURL.attributeTypeMap;
		}
	}